+++ CHANGELOG.md
@@ -5,6 +5,20 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+
+- Added `check` module. This module contains the same functions as the
+  `assert` module, but they return a boolean and mark the test as failed
+  instead of failing the test immediately.
+
+- `NotErrorIs`, `NotErrorContains`, and `NotDeepEqual` functions.
+
+### Removed
+
+- `LenEqual`, `Assert`, and `Check` functions.
+
 ## [0.2.0] - 2024-01-28
 
 ### Changed